Add feature to retrieve snippets from files
Each snippet is delimited by two markers: `#dd/start` and `#dd/end`.
`#dd/start` must be followed by a name, and this name should be unique
in this project.

e.g: in <project-root>/src/test.clj

    (some code
      (calling other code)
      ;; #dd/start my-snippet
      (more calls)
      (some more
         (calls to other functions))
      ;; some comment
      (even-more calls)
      ;; #dd/end
      (this is over folks))

will result in the following inside docudata.edn

    {:snippets
      {:my-snippet
        {:content "(more calls)\n(some more\n  (calls to other functions))\n;; some comment\n(even-more calls)"
         :file "src/test.clj"}}

Notice that the content of the snippet is left trimmed while
preserving relative indentation.

Snippets can be found in any file in the project.